fix warning in su_get_module_path()

+++ src/proc.c
@@ -19,6 +19,7 @@
 #define _GNU_SOURCE
 #include <dlfcn.h>
 #include "proc.h"
+#include <stdint.h>
 
 /********************************************************************/
 
@@ -27,7 +28,7 @@ const char * su_get_module_path(void * p)
     Dl_info  info;
     if (dladdr(p, &info) != 0)
     {
-        g_print("pointer 0x%llX identified as module %s\n", (unsigned long long)p, info.dli_fname);
+        g_print("pointer 0x%llX identified as module %s\n", (unsigned long long)(uintptr_t)p, info.dli_fname);
         return info.dli_fname;
     }
     return NULL;
